Add test to ApiSelection component. Extract dropdown mock to testUtils.
