+++ src/client/src/containers/CosmosDbModal/APISelection/ApiSelection.test.tsx
@@ -1,0 +1,42 @@
+import * as React from "react";
+import "@testing-library/jest-dom";
+import ApiSelection from ".";
+import { RenderResult, fireEvent } from "@testing-library/react";
+import "@testing-library/jest-dom/extend-expect";
+import { renderWithIntl } from "../../../testUtils";
+import { COSMOS_APIS } from "../../../utils/constants";
+
+jest.mock("../../../components/Dropdown", () => require("../../../testUtils").dropdownMock);
+
+describe("ApiSelection", () => {
+  let props: any;
+  let wrapper: RenderResult;
+
+  beforeEach(() => {
+    props = {
+      initialApi: "",
+      onApiChange: jest.fn(),
+    };
+  });
+
+  it("renders without crashing", () => {
+    wrapper = renderWithIntl(<ApiSelection {...props} />);
+    expect(wrapper).toBeDefined();
+  });
+
+  it("If has initial Api, onApiChange notify selected Api", () => {
+    props.initialApi = COSMOS_APIS.MONGO;
+    wrapper = renderWithIntl(<ApiSelection {...props} />);
+    expect(props.onApiChange).toHaveBeenCalledTimes(1);
+    expect(props.onApiChange).toHaveBeenCalledWith(COSMOS_APIS.MONGO);
+  });
+
+  it("When selected an Api in a dropdown, onApiChange notify selected Api", () => {
+    const selectedApi = COSMOS_APIS.SQL;
+    wrapper = renderWithIntl(<ApiSelection {...props} />);
+    const dropdown = wrapper.getByTestId("dropdown");
+    fireEvent.change(dropdown, { target: { label: selectedApi, value: selectedApi } });
+    expect(props.onApiChange).toHaveBeenCalledTimes(1);
+    expect(props.onApiChange).toHaveBeenCalledWith(selectedApi);
+  });
+});
