+++ src/client/src/testUtils.tsx
@@ -2,6 +2,7 @@ import * as React from "react";
 import { Provider } from "react-redux";
 import { IntlProvider } from "react-intl";
 import { render } from "@testing-library/react";
+import { IDropdownProps } from "./components/Dropdown";
 
 const renderWithIntl = (component: any) => {
   return render(<IntlProvider locale="en">{component}</IntlProvider>);
@@ -11,4 +12,22 @@ const renderWithStore = (component: any, store: any) => {
   return renderWithIntl(<Provider store={store}>{component}</Provider>);
 };
 
-export { renderWithIntl, renderWithStore };
+const dropdownMock = ({ options, value, handleChange }: IDropdownProps) => {
+  const handleInputChange = (event: any) => {
+    if (handleChange) {
+      handleChange(event.currentTarget);
+    }
+  };
+  const newValue = value ? value.label : undefined;
+  return (
+    <select data-testid="dropdown" value={newValue} onChange={handleInputChange}>
+      {options.map(({ label, value }: any) => (
+        <option key={value} value={value}>
+          {label}
+        </option>
+      ))}
+    </select>
+  );
+}
+
+export { renderWithIntl, renderWithStore, dropdownMock };
